\name{acacia_pinus}
\alias{acacia_pinus}
\docType{data}
\title{
Data for Acacia and Pinus tree distributions
}
\description{
Data for occurrences of Acacia and Pinus species based on geographic observations.
}
\usage{data(acacia_pinus)}
\format{
  A \code{data.frame} with 37845 observations on the following 3 variables.
  \describe{
    \item{\code{Species}}{a \code{character} containing 139 unique values}
    \item{\code{Latitude}}{a numeric vector}
    \item{\code{Longitude}}{a numeric vector}
  }
}

\source{
Occurrence data come from the BIEN database (\code{https://biendata.org/}).
}
\references{
Blonder, B., Lamanna, C., Violle, C., Enquist, B. The n-dimensional hypervolume. Global Ecology and Biogeography (2014).
}
