\name{circles}
\alias{circles}
\docType{data}
\title{
Circles simulated dataset
}
\description{
Data generated by picking 100 points randomly within a circle of radius 1. See the vignette on occupancy for information about its usage.
}
\usage{data(circles)}
\format{
  A \code{list} with 20 objects of class \code{matrix}. Each \code{matrix} contains 100 rows and 2 columns.
  \describe{
    \item{\code{[,1]}}{x coordinate}
    \item{\code{[,2]}}{y coordinate}
  }
}
