
\name{get_volume}
\alias{get_volume}
\alias{get_volume.Hypervolume}
\alias{get_volume.HypervolumeList}
\alias{get_volume,Hypervolume-method}
\alias{get_volume,HypervolumeList-method}



%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Extract volume
}
\description{
Extract volume from Hypervolume or HypervolumeList object
}
\usage{
\method{get_volume}{Hypervolume}(object)
\method{get_volume}{HypervolumeList}(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
  A \code{Hypervolume} or \code{HypervolumeList} object
}
}
\value{
A named numeric vector with the volume of each input hypervolume
}

