\name{hypervolume_redundancy}
\alias{hypervolume_redundancy}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Redundancy of a point in a hypervolume
}
\description{
Estimates squared probability density at a given point. This metric is proportional to the number of data points multiplied by the probability density at a point.
}
\usage{
hypervolume_redundancy(...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{\dots}{
Arguments to be passed to \code{hypervolume_estimate_probability}
}
}

\seealso{
\code{\link{hypervolume_estimate_probability}}
}