\name{HypervolumeList-class}
\Rdversion{1.1}
\docType{class}
\alias{HypervolumeList-class}

\title{Class HypervolumeList}
\description{
A convenience class used to streamline plotting and other hypervolume output.
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("HypervolumeList", ...)}.
}
\section{Slots}{
  \describe{
    \item{\code{HVList}:}{Object of class \code{"list"}. A list of hypervolumes. }
  }
}