\name{print.Hypervolume}
\alias{print.Hypervolume}
\alias{print.HypervolumeList}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print summary of hypervolume
}
\description{
Summarizes all slots of \code{\link{Hypervolume-class}} object.
}
\usage{
\method{print}{Hypervolume}(x, ...)
\method{print}{HypervolumeList}(x, ...)
}
\arguments{
  \item{x}{
The hypervolume to summarize
}
  \item{...}{
 
}
}
\value{
None; used for the side-effect of printing.
}
