\name{summary.Hypervolume}
\alias{summary.Hypervolume}
\alias{summary.HypervolumeList}
\alias{show.Hypervolume}
\alias{show.HypervolumeList}
\title{
Summary of hypervolume
}
\description{
Prints basic information about Hypervolume or HypervolumeList structure.
}
\usage{
\method{summary}{Hypervolume}(object, ...)
\method{summary}{HypervolumeList}(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
The hypervolume to summarize
}
  \item{...}{
 
}
}
\value{
None; used for the side-effect of printing.
}