\name{hypervolume-package}
\alias{hypervolume-package}
\docType{package}
\title{
N-dimensional hypervolume operations
}
\description{
Estimates the shape and volume of high-dimensional objects and performs set operations: intersection / overlap, union, unique components, and inclusion test. Can measure the n-dimensional ecological hypervolume and perform species distribution modeling.
}

\details{
A FAQ can be found at http://www.benjaminblonder.org/hypervolume_faq.html.



}

\author{
Benjamin Blonder <bblonder@gmail.com>
}
\references{
Blonder, B., Lamanna, C., Violle, C. and Enquist, B. J. (2014), The n-dimensional hypervolume. Global Ecology and Biogeography, 23: 595-609. doi: 10.1111/geb.12146
}