\name{hypervolume-package}
\alias{hypervolume-package}
\docType{package}
\title{
\packageTitle{hypervolume}
}
\description{
\packageDescription{hypervolume}
}
\details{
A frequently asked questions document (FAQ) can be found at http://www.benjaminblonder.org/hypervolume_faq.html. More details are also available in a user guide within our 2018 paper (see reference below).

}
\author{
\packageAuthor{hypervolume}

Maintainer: \packageMaintainer{hypervolume}
}
\references{
Blonder, B., Lamanna, C., Violle, C. and Enquist, B. J. (2014), The n-dimensional hypervolume. Global Ecology and Biogeography, 23: 595-609. doi: 10.1111/geb.12146

Blonder, B. Do Hypervolumes Have Holes?, The American Naturalist, 187(4) E93-E105. doi: 10.1086/685444

Blonder, B., Morrow, C.B., Maitner, B., et al. New approaches for delineating n-dimensional hypervolumes. Methods Ecol Evol. 2018;9:305-319. doi: 10.1111/2041-210X.12865
}

