\name{HypervolumeList-class}
\Rdversion{1.1}
\docType{class}
\alias{HypervolumeList-class}

\title{Class \code{"HypervolumeList"}}
\description{
A class used for storing more than one hypervolume.
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("HypervolumeList", ...)}.
%%  ~~ describe objects here ~~ 
}
\section{Slots}{
  \describe{
    \item{\code{HVList}:}{Object of class \code{"list"} containing multiple hypervolumes}
  }
}
\keyword{classes}
